+++ src/Kit/Layout/Section/types.d.ts
@@ -1,0 +1,8 @@
+import {KitColorProp} from '@utils/functions/types';
+import {CSSProperties} from 'react';
+
+interface IKitSection {
+    className?: string;
+    style?: CSSProperties;
+    borderColor?: KitColorProp;
+}