+++ stories/02-Design System/02-Layout/Section/exemples/index.tsx
@@ -1,0 +1,15 @@
+import Basic from './basic';
+import Colors from './colors';
+
+import BasicSource from './basic?raw';
+import ColorsSource from './colors?raw';
+
+export const Sources = {
+    Basic: BasicSource,
+    Colors: ColorsSource
+};
+
+export default {
+    Basic,
+    Colors
+};